execution: follow up tidy bench guideline
